ready for midterm

+++ app/api/api_v2/endpoints/message_api.py
@@ -1,15 +1,13 @@
-from fastapi import APIRouter, Depends, HTTPException, Query, Path
-from fastapi.security import HTTPAuthorizationCredentials
-from starlette.responses import JSONResponse
-from starlette import status
+from http.client import HTTPException
+
+from fastapi import APIRouter, Depends, Query
 from app.db.db import get_session
 from sqlalchemy.ext.asyncio.session import AsyncSession
-#from app.services.chat_service import ChatService
+
 from typing import Annotated
 import logging
 from app.models.message_model import *
 from app.services.auth_service import auth_wrapper, auth_wrapper_lamda
-from starlette.requests import Request
 from fastapi.responses import StreamingResponse
 
 from app.services.chat_service import ChatServiceV2
@@ -28,7 +26,7 @@ async def get_history(
     #user =auth_wrapper(3)  # Passing chatId to the auth wrapper
 
     if user == 0:
-        raise HTTPException(status_code=401, detail="Invalid token")
+        return "Require token to access bot!"
 
     print("GET HISTORY OF ", chatid)
     history = await ChatServiceV2.get_chat_history(chatid)
@@ -40,11 +38,12 @@ async def send_message(
                     message: MessageCreate,
                     #chat_service: ChatService = Depends(),
                     session: AsyncSession = Depends(get_session),
-                    #user=Depends(auth_wrapper)
+                    user=Depends(auth_wrapper)
                     ):
     try:
-        # if user == 0:
-        #     raise HTTPException(status_code=401, detail="Invalid token")
+        if user == 0:
+            return "Require token to access bot!"
+
         response_generator, full_bot_response, message_id = await ChatServiceV2.send_message(message)
         async def streaming_response():
             async for chunk in response_generator:
@@ -52,24 +51,24 @@ async def send_message(
 
             # Once the response is finished, save the full response to the database
             bot_message_content = ''.join(full_bot_response)
+            print("BOT_MESSAGE_CONTENT: ", bot_message_content)
             if bot_message_content:
                 await ChatServiceV2.update_bot_message(message_id,message.chatId, bot_message_content)
                 # Save the bot's message to the database
         # Return the streamed response to the client
         return StreamingResponse(streaming_response(), media_type='text/plain')
     except Exception as e:
-        logger.error(e)
-        return "Something went wrong"
+        return "Xin lỗi nhưng bạn có thể đặt lại câu hỏi được không ạ?"
 
 @router.delete("/chat")
 async def delete_message(
                     chatid: Annotated[int | None, Query()]=None,
                     #chat_service: ChatService = Depends(),
                     #session: AsyncSession = Depends(get_session),
-                    #user=Depends(auth_wrapper)
+                    user=Depends(auth_wrapper)
                     ):
-    # if user == 'fail':
-    #     raise HTTPException(status_code=401, detail="Invalid token")
+    if user == 0:
+        raise HTTPException(status_code=401, detail="Invalid token")
 
     print("DELETE MESSAGE FROM ", chatid)
     result = await ChatServiceV2.delete_message(chatid)
